% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{.estimate_dens}
\alias{.estimate_dens}
\title{Density wrapper.}
\usage{
.estimate_dens(x, dist, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{dist}{Density function or logspline object.}

\item{...}{Arguments passed onto density function.}
}
\value{
Numeric.
}
\description{
Density wrapper.
}
\author{
Stephen R. Martin
}
\keyword{internal}
