% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{.estimate_dist}
\alias{.estimate_dist}
\title{Estimates (logspline) density of vector.}
\usage{
.estimate_dist(x, lb = -1, ub = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{lb}{Numeric. Lower bound.}

\item{ub}{Numeric. Upper bound.}
}
\value{
Logspline object.
}
\description{
Estimates (logspline) density of vector.
}
\author{
Stephen R. Martin
}
\keyword{internal}
