% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%IfNull\%}
\alias{\%IfNull\%}
\title{Helper for defaulting NULL values.}
\usage{
x \%IfNull\% expr
}
\arguments{
\item{x}{Object.}

\item{expr}{Expression.}
}
\value{

}
\description{
Default setting helper.
}
\details{
If x is null, replace with expr.
}
\author{
Stephen Martin
}
\keyword{internal}
