% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.corCompare}
\alias{print.summary.corCompare}
\title{Print method for summary.corCompare}
\usage{
\method{print}{summary.corCompare}(x, ...)
}
\arguments{
\item{x}{summary.corCompare object.}

\item{...}{Not used.}
}
\value{
x (Invisibly).
}
\description{
Print method for summary.corCompare
}
\author{
Stephen R. Martin
}
