% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.corCompare}
\alias{summary.corCompare}
\title{Summary method for corCompare objects.}
\usage{
\method{summary}{corCompare}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{corCompare object.}

\item{prob}{Numeric (0, 1). Amount of probability mass to contain in the credible interval.}

\item{...}{Not used.}
}
\value{
summary.corCompare object. List containing the summary tables (summary) and meta-data (meta).
}
\description{
Summary method for corCompare objects.
}
\author{
Stephen R. Martin
}
