% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{tbeta}
\alias{tbeta}
\alias{dtbeta}
\alias{qtbeta}
\alias{ptbeta}
\alias{rtbeta}
\alias{dtbeta_diff}
\alias{ptbeta_diff}
\alias{qtbeta_diff}
\title{Transformed beta distribution}
\usage{
dtbeta(x, mu, psi, log = FALSE)

qtbeta(p, mu, psi)

ptbeta(q, mu, psi)

rtbeta(n, mu, psi)

dtbeta_diff(x, mus, psis, log = FALSE)

ptbeta_diff(q, mus, psis, lower.tail = TRUE)

qtbeta_diff(p, mus, psis)
}
\arguments{
\item{x}{Numeric vector. (-1, 1) for tbeta, and (-2, 2) for tbeta_diff.}

\item{mu}{Numeric vector (-1, 1). Mean parameter.}

\item{psi}{Numeric vector (0, Inf). Shape parameter.}

\item{log}{Logical (Default: FALSE). Whether to take the log of the probability.}

\item{p}{Numeric vector. Probabilities.}

\item{q}{Numeric vector. Quantiles.}

\item{n}{Integer. Number of random samples.}

\item{mus}{Numeric vector (Length 2). Defined in (-1, 1). Mu parameters for the differenced beta variates.}

\item{psis}{Numeric vector (length 2). Defined in (0, Inf). Psi parameters for the differenced beta variates.}
}
\description{
Density, distribution function, quantile function, and random generation for the transformed beta distribution with mean equal to mu and shape parameter psi.
}
\details{
The transformed beta distribution is the beta distribution, shifted and scaled to be defined between -1 and 1.
The tbeta distribution is therefore useful for distributions on individual correlations.
Importantly, it is not useful as a distribution for elements within a correlation matrix beyond 2x2, as it has no guarantee that the matrix is positive definite.

\deqn{E(x) = mu}
\deqn{V(x) = \frac{(1 + \mu)(1 - \mu)}{\psi + 1}}

The tbeta distribution is a rescaled, shifted beta distribution under the mean-shape parameterization.
\deqn{p(x|mu, psi) = \frac{(.5(x + 1))^{.5(\mu + 1)\psi - 1}(1 - .5(x + 1))^{\psi(1 - .5(\mu + 1)) - 1}}{2\Beta(.5(\mu + 1)\psi, \psi(1 - .5(\mu + 1)))}}

Or more simply: It is a beta distribution, shifted by -.5, and multiplied by 2.
The density is therefore equivalent to inverse-transforming the variate and mu, and solving for the Beta alpha and beta parameters.
The jacobian of this adjustment is .5.

The tbeta_diff distribution is the distribution of differences in tbeta-distributed variables.
That is, if X ~ tbeta(mu_x, psi_x) and Y ~ tbeta(mu_y, psi_y), then (X - Y) ~ tbeta_diff(mu_x, mu_y, psi_x, psi_y).
E(X - Y) = mu_x - mu_y
V(X - Y) = V(X) + V(Y), where V() is defined above.
}
\author{
Stephen R. Martin
}
