% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corCompare.R
\name{cor_compare}
\alias{cor_compare}
\title{Fit cor_compare model.}
\usage{
cor_compare(formula, data, ...)
}
\arguments{
\item{formula}{Formula. See details.}

\item{data}{Data.frame.}

\item{...}{Options to pass onto sampling.}
}
\value{
Stanfit.
}
\description{
Fits the correlation-compare model.
}
\details{
The formula is multipart: y ~ x | groups.
I.e., outcome ~ predictor | between + within + between2

The model effectively estimates a covariance/correlation between outcome and predictor, for each possible combination of the specified grouping variables.
}
\author{
Stephen R. Martin
}
