% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dtbeta}
\alias{dtbeta}
\alias{ptbeta}
\alias{rtbeta}
\alias{dtbeta_diff}
\title{Transformed Beta distribution.}
\usage{
dtbeta(x, mu, psi, log = FALSE)

ptbeta(q, mu, psi)

rtbeta(n, mu, psi)

dtbeta_diff(x, mus, psis, log = FALSE)
}
\arguments{
\item{x}{Numeric vector \link{-2, 2}.}

\item{mu}{Numeric vector \link{-1, 1}. Mean parameter.}

\item{psi}{Numeric vector \link{0, Inf}. Shape parameter.}

\item{log}{Logical (Default: FALSE). Whether to return densities on log scale.}

\item{q}{Numeric vector. Quantiles for the tbeta distribution.}

\item{n}{Number of samples.}

\item{mus}{Numeric vector (length 2). Mu-parameters to two tbeta distributions.}

\item{psis}{Numeric vector (length 2). Psi-parameters to two tbeta distributions.}
}
\value{
Numeric vector. Densities

Numeric vector.

Numeric vector.

Numeric vector.
}
\description{
Density and sampling function for transformed beta distribution.

Implied PDF of difference in two tbeta random variates.
}
\details{
The transformed beta distribution is a mean-scale shifted beta distribution ranging from -1 to 1.
The tbeta distribution is useful as a prior for a correlation.
}
\author{
Stephen R. Martin

Stephen R. Martin

Stephen R. Martin

Stephen R. Martin
}
